requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
dataclasses-json==0.6.7
deprecation==2.1.0
exceptiongroup==1.3.0
frozenlist==1.5.0
gotrue==2.9.2
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langsmith==0.1.147
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.1.0
numpy==1.24.4
ollama==0.4.8
orjson==3.10.15
packaging==24.2
postgrest==0.16.11
propcache==0.2.0
pydantic==2.10.6
pydantic-core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
realtime==1.0.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
storage3==0.7.7
StrEnum==0.4.15
supabase==2.6.0
supafunc==0.5.1
tenacity==8.5.0
typing-extensions==4.13.2
typing-inspect==0.9.0
urllib3==2.2.3
websockets==12.0
yarl==1.15.2
